+++ src/main/java/lotto/domain/LottoTickets.java
@@ -5,6 +5,7 @@ import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
+import lotto.utils.NumberGenerator;
 
 public class LottoTickets {
