+++ src/main/java/lotto/utils/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.utils;
 
 import java.util.Collections;
 import java.util.List;
