+++ src/test/java/lotto/domain/LottoTicketsTest.java
@@ -4,6 +4,8 @@ import static org.assertj.core.api.Assertions.*;
 
 import java.util.List;
 import lotto.TestLottoFactory;
+import lotto.utils.NumberGenerator;
+import lotto.utils.RandomNumberGenerator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
